requirements.txt
matplotlib
numpy
torch
torchvision
pyyaml
ipykernel